first day of refactoring. did some good stuff in functions and css
