+++ r-p-s.js
@@ -1,5 +1,5 @@
 import resultChecker from './result-checker.js';
-
+import getRandomThrow from './get-random-throw.js';
 
 const throwButton = document.getElementById('throwButton');
 const winSpan = document.getElementById('win-span');
@@ -11,14 +11,6 @@ let totalLosses = 0;
 let totalDraws = 0;
 let winLossPct = 0;
 
-const getRandomThrow = (computerThrow) => {
-    if (computerThrow === 0)
-        return 'rock';
-    if (computerThrow === 1)
-        return 'paper';
-    if (computerThrow === 2)
-        return 'scissors';
-};
 
 const roshambeau = () => {
     let computerThrow = Math.floor(Math.random() * 3);
@@ -36,9 +28,7 @@ const updateSpans = () => {
     drawSpan.textContent = totalDraws;
 };
   
-
-
-throwButton.addEventListener('click', () => {
+const trueWinner = () => {
 
     const declaredWinner = roshambeau();
     if (declaredWinner === 'win') {
@@ -49,7 +39,7 @@ throwButton.addEventListener('click', () => {
     
     winLossPct = Math.round(totalWins / (totalDraws + totalLosses + totalWins) * 100);
     updateSpans();
+};
 
 
-
-});
+throwButton.addEventListener('click', trueWinner);
